+++ lib/oneview-sdk/resource.rb
@@ -327,12 +327,11 @@ module OneviewSDK
     # @param [Array<OneviewSDK::Resource>] resources The list of resources
     # @return [Array<String>] List of uris
     # @raise IncompleteResource if 'uri' is not set for each resource.
-    def get_and_ensure_uri_for(resources)
-      uris = resources.map do |resource|
+    def get_and_ensure_uri(resources)
+      resources.map do |resource|
         resource.ensure_uri
         resource['uri']
       end
-      uris
     end
 
     # Fail for methods that are not available for one resource